Fix bug preventing the setting of mulltiple dependant tasks
When setting a task to have many dependants, we add each depended task to
the children of the dependant task, and then remove the depended task
from the list of active tasks. This act of removing a depended task
changes the indexes of other tasks in the list.

To get around this, we build a list of tasks to removed, sort this list
by descending index, and then remove them one by one.