Basically a complete rewrite
Redesigned the database. Previously we stored all the visible tasks, and
all tasks that were hidden were children of those tasks.

Now the visible tasks are maintained separately, and there is a single
flat map of id -> task for all tasks in the system.

The children field on tasks is the tasks that must be completed before
the task is visible.